repos/isaias-goncalves/requirements.txt
yfinance
streamlit
pandas
python-dotenv
openai
google-generativeai
pytest
pytest-mock
requests
beautifulsoup4
lxml
python-dateutil
curl_cffi
typing-extensions
